requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
openai-whisper
ffmpeg-python
celery
redis